+++ lib/evil_events/core/events/serializers/message_pack/config.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+class EvilEvents::Core::Events::Serializers
+  class MessagePack
+    # @api private
+    # @since 0.4.0
+    class Config < Base::GenericConfig
+      configure do
+        setting :engine, reader: true
+
+        setting :mpacker, reader: true do
+          setting :configurator, ->(engine) {}
+        end
+      end
+    end
+  end
+end