+++ CHANGELOG.md
@@ -25,22 +25,54 @@ All notable changes to this project will be documented in this file.
   - `YourEventClass.emit!(adapter: :your_adapter_identifier, **event_attrs)`;
   - `your_event.emit!(adapter: :your_adapter_identifier)`;
 - Added XML event serialization format (no more JRuby at this moment **:(**):
-  - `EvilEvents::Serializer.load_from_xml(event)` - returns an event object;
+  - `EvilEvents::Serializer.load_from_xml(xml)` - returns an event object;
   - `your_event.serialize_to_xml` - returns xml string;
   - `your_event.dump_to_xml` - returns xml string (`serialize_to_xml` alias);
+- Added MessagePack event serialization format (no more JRuby at this moment **:(**):
+  - `EvilEvents::Serializer.load_from_msgpack(message)` - returns an event object;
+  - `your_event.serialize_to_msgpack` - returns messagepack string;
+  - `your_event.dump_to_msgpack` - returns messagepack string (`serialize_to_msgpack` alias);
 - Added an ability to check an event object similarity:
   - `your_event.similar_to?(another_event)` - `returns` true if `another_event` has equal id/type/metadata/paylaod
     attributes (and interface) - otherwise returns `false`.
 
 ### Changed
-- Removed `EvilEvents::CombinedContext` class and submodules => `Symbiont` is used instead (gem `symbiont-ruby`);
-- Added specific serialization error classes (inherited from `EvilEvents::SerializationError`:
-  - `EvilEvents::HashSerializationError`;
-  - `EvilEvents::HashDeserializationError`;
-  - `EvilEvents::JSONSerializationError`;
-  - `EvilEvents::JSONSerializationError`;
-  - `EvilEvents::XMLSerializationError`;
-  - `EvilEvents::XMLSerializationError`.
+- Removed `EvilEvents::Shared::Configurable` => `EvilEvents::Shared::AnyConfig` is used instead (smart dry-configurable wrapper);
+- Removed `EvilEvents::Shared::CombinedContext` class and submodules => `Symbiont` is used instead (gem `symbiont-ruby`);
+- Refactored event serialization abstraction:
+  - Each event serializer represents a complex logic now and is constructed by a set of objects:
+    - engines - low-level data-transformers, works with event serialization state (event data mapper);
+    - packer - responsible for serialization via corresponding engine  (works with events and event serialization state);
+    - unpacker - responsible for deserialization logic via corresponding engine (works with low-level data and event serialization state);
+    - serializers container;
+    - config - configuration point of the each serializer;
+    - event serialization state - event data mapper (id, type, payload, metadata), engines works with this state;
+  - Supported engines:
+    - Hash (`:native`) - based on ::Hash class;
+    - JSON (`:native`) - based on ::JSON class;
+    - MessagePack (`:mpacker`) - based on MessagePack gem;
+    - XML (`:ox`) - based on Ox gem;
+  - Added configuration points for the each type of serializer:
+    - `EvilEvents::Config.options.serializers` - root
+    - `EvilEvents::Config.options.serializers.json.engine` - `:native` by default;
+    - `EvilEvents::Config.options.serializers.hashing.enigne` - `:native` by default;
+    - `EvilEvents::Config.options.serializers.xml.engine` - `:native` by default;
+    - `EvilEvents::Config.options.serializers.msgpack.engine` - `:mpacker` by default;
+    - `EvilEvents::Config.options.serializers.msgpack.mpacker.configurator` - `->(engine) {}` by default;
+- Added specific serialization error classes (inherited from `EvilEvents::SerializationError`/`EvilEvents::DeserializationError` respectively):
+  - `EvilEvents::SerializersError`
+  - `EvilEvents::SerializationError`
+  - `EvilEvents::JSONSerializationError`
+  - `EvilEvents::XMLSerializationError`
+  - `EvilEvents::HashSerializationError`
+  - `EvilEvents::MessagePackSerializationError`
+  - `EvilEvents::DeserializationError`
+  - `EvilEvents::JSONDeserializationError`
+  - `EvilEvents::XMLDeserializationError`
+  - `EvilEvents::HashDeserializationError`
+  - `EvilEvents::MessagePackDeserializationError`
+  - `EvilEvents::SerializationEngineError`
+  - `EvilEvents::UnrecognizedSerializationEngineError`
 
 ## [0.3.1] - 2018-03-01
 ### Fixed